Examples/gRPC-protobuf/_requirements.txt
Flask==3.1.0
grpcio==1.71.0
protobuf==6.30.2
